# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Birthday Message",
        "description": "Birthday Message API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Display a listing of the resource.",
                "description": "User",
                "operationId": "user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "John"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Doe"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "bourke street"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "melbourne"
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "australia"
                                                    },
                                                    "birth_date": {
                                                        "type": "string",
                                                        "example": "2024-01-24T17:00:00.000000Z"
                                                    },
                                                    "timezone": {
                                                        "type": "string",
                                                        "example": "Australia/Melbourne"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john.doe@mail.com"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-01-25T04:41:09.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-01-25T04:41:09.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "User Store",
                "operationId": "userStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "address",
                                    "city",
                                    "country",
                                    "birth_date"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "bourke street"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "melbourne"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "australia"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "2024-01-24"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "Australia/Melbourne"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "bourke street"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "melbourne"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "australia"
                                                },
                                                "birth_date": {
                                                    "type": "string",
                                                    "example": "2024-01-24T17:00:00.000000Z"
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "Australia/Melbourne"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@mail.com"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-01-25T04:41:09.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-01-25T04:41:09.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 2002
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "The last name field is required."
                                                },
                                                "error": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The last name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}